feat: update logging to print string error codes for gripper
